Add/remove some MinWise constructors.
Remove the constructor that initialized a new MinWise instance
from a set of string integer (e.g., "123") data.  This was nothing
other than a for loop basically.

Add the constructor NewMinWiseFromSignature to convert from []uint64
to a MinWise instance.  This allows for further streaming when loading
a signature from saved data, provided the original hash functions used
to create the signature in the first place are known.