Tweak MinWise methods to implement MinHash.
Encountered an issue when defining an interface whose defining
methods accept that interface as a parameter.  For example,
one the Similarity method of the MinHash interface is currently
of type func(Minhash) float64.  The MinWise similarity method
was of type func(*MinWise) float64.  This caused a *MinWise instance
to not actually implement the MinHash interface.  So we changed the
relevant MinWise methods to accept MinHash interfaces.  This creates
some other issues such as not being able to directly access
the input's minimums attribute, but this shouldn't pose a problem.